dpm-solver/score_sde_pytorch/requirements.txt
ml-collections==0.1.1
tensorflow-gan==2.1.0
tensorflow_io==0.22.0
tensorflow_datasets==4.4.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorboard==2.8.0
tensorflow-probability==0.15.0
absl-py==1.4.0
torch>=1.7.0
torchvision
ninja==1.11.1.1
numpy==1.24.2
jax==0.4.8
jaxlib==0.4.7
pillow==10.2.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io==0.32.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.19.0
tf-keras==2.15.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.66.2
triton==2.0.0
